backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
mysqlclient==2.2.0
celery==5.3.4
redis==5.0.1
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3
Pillow==10.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
